DM Load All Character Sheets On Start
When a DM starts they will load all the characters sheets into hidden iframes. This will populate the character/token info automatically. When the DM clicks on a player sheet, the iframe containing that character's sheet will be shown.
If a player opens and closes their sheet on their end, on the DM's side the iframe containing that player's sheet will be marked as changed (iframe attribute data-changed:true), and the iframe will reload when the DM opens it.
Added a "reload" (🗘) button to the top left of the character sheet which will refresh the iframe containing the sheet.
Character sheets now act the same way for DMs and players, with the "SHEET" button both opening and closing the respective sheets, having close/reload buttons for both DMs and players, and both having opening/closing animations.