lock/unlock/player_sheet_closed
Changed locking/unlocking to work with changes from the DM pre-loading characters. Added player_sheet_closed message, which marks the character's sheet iframe to be reloaded the next time it is opened by the DM.
Use the ID from the end of the character sheet url for lock/unlock/player_sheet_closed, since the url can be different depending on if the user is a DM or not.