Message Broker confirmations
Added a unique connection_id to the message broker (MB).
Added onlineUserList to the MB, an array of connection IDs of each other online user.
Added sent_messages to the MB, and array of the IDs of all sent messages.
Added received_messages to the MB, and array of the IDs of all received messages.
Added message_confirmations, a dictionary where each key is the ID of a messages sent and each value is and arrays of the connection ids of all users who have send confirmations to that message.
Added unconfirmed_messages, a dictionary where each key is the ID of a message sent but not yet confirmed by everyone in the online user list

Added the ability to send and receive message bundles, which are messages where the data is an array of messages.

When handling messages, the message ID will be checked against the received_messages list, and if it is there it will not be handled (but a confirmation will still be sent for it).

Every 10 seconds MB send an "AmOnline" message to everyone else with their connection id. The receivers add them to their online users list.

When a user sends a message, the id is added to sent_messages, a new entry with an empty array is added to message_confirmations, and (if a confirmation is required) the message is added to unconfirmed_messages.

MB sends a confirmation for every message received (except confirmations, AmOnline, and message bundle messages). These confirmations include the original message id and the MB's connection id. When a confirmation is received, the connection id from the confirmation is added to the array of the message in message_confirmations.

Every second the MB will loop through the messages in unconfirmed messages. If messages have received confirmations, remove them from the unconfirmed messages list. If any unconfirmed messages are older than 1000ms and younger than 10,000ms and have not been confirmed by all online users, resend them in a message bundle. If the message has been confirmed by everyone on the online user list, remove them from the unconfirmed messages list. If the message age is older than 10,000ms, remove the message from the unconfirmed messages list.