Fix: Get Super Mega Wizard to appear again when clicking the button. (#420)
* Fix for super mega wizard see details

Stop check for if not player token on a token with an undefined id. Still run the code that should run since it's not a player. This was causing an error running .isPlayer on undefined when clicking the super mega wizard.

* Fix for super mega wizard see details

After speaking with mikedave he suggested it best to have the function checking if it's a player return false if undefined. I think this is a much better solution.

* Clean up for merge

* Fix code spacing

Oh god