Add conditions to quick roll, and UI changes
I added conditions to the quick roll menu. I also made some UI updates based on some feedback I saw.

There was an issue with the dropdowns when the menu is a popout. I didn't have time to look into very much but the dropdowns don't seem to work on the cloned window. It didn't feel terribly important, but if someone who knows JS better than me can take a crack at it.