refactor: improve language selector design a bit
Previous implementation used position absolute to place the button,
did not reset the background so on firefox it would stand out
and the button was very narrow.

Now it's slightly nicer both in code and in visuals, I hope.

I would actually prefer this selector to be the same as the mui one on
the top of the game log, but I didn't find a way to do that

I'm also not quite sure that I changed popoutGamelogCleanup correctly